requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.8
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
